#210 Change BroadbandPacket.php to InternetPackage.php

+++ database/migrations/2024_09_09_084737_create_internet_packages.php
@@ -19,7 +19,7 @@ return new class extends Migration {
                 ->cascadeOnUpdate();
             $table->string('name');
             $table->integer('capacity');
-            $table->double('price');
+            $table->decimal('price', 15, 4);
             $table->timestamps();
         });
     }
